rev-parse: deprecate use as an option sifter
This command used to be used to distinguish between parameters meant
for rev-list and others, but this is deprectated now, and one should not
rely on rev-parse keeping up to date with rev-list. State this in the
man page and reword the title according to the primary use (converting
symbolic names to SHA1's).

Reference: <http://permalink.gmane.org/gmane.comp.version-control.git/146118>

Signed-off-by: Michael J Gruber <git@drmicha.warpmail.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>