perl libs: perl -w + use warnings is redundant
Change the Perl scripts to turn on lexical warnings, but only in Perl
scripts that definitely require 5.6.0 or later. I.e. those that either
did 'use warnings' in addition to -w already, or used the three-arg
form of open.

-w sets warnings for all code that interpreter runs. "use warnings" is
lexically scoped. The former is probably not what the authors wanted.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>