gettext: Add a Gettext interface for Perl
Make Git's gettext messages available to Perl programs through
Locale::Messages. Gracefully fall back to English on systems that
don't contain the module.

This contains some makefile hacks to pass localedir to
perl/Makefile. Which in turn passes it to perl.mak, which'll use it to
search/replace occurances of ++LOCALEDIR++ in .pm files.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>