Makefile: Don't install Gettext .mo files if NO_GETTEXT
Change the Makefile not to build/install .mo files as part of the
default target if NO_GETTEXT=1 is given.

This was both redundant, and meant that the Perl and Shell programs
would use the Gettext strings, since they don't use the
git_setup_gettext() function in gettext.c, which is compiled to a stub
if NO_GETTEXT=1 is set.

Reported-by: Jakub Narebski <jnareb@gmail.com>
Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>