Add infrastructure for translating Git with gettext
Change the build process to use GNU's libintl if it's available. If
not we define our own skeleton replacement functions which degrade
gracefully to English.

Gettext .mo files will be installed and looked for in
$(prefix)/share/locale. GIT_LOCALEDIR can also be set to override
that, this variable is only intended for testing Git itself.

Based on work by Jeff Epler <jepler@unpythonic.net>.