Add localization support to Git with gettext
All of the interface messages in Git core are currently hardcoded to
English. This patch changes that by optionally enabling translation of
the core C, Shell and Perl programs via GNU gettext. If you set the
appropriate LC_* variables Git will speak your language.

If gettext isn't available, or if Git is compiled with
NO_GETTEXT=YesPlease, then Git fall back on its previous behavior of
only speaking English.

With NO_GETTEXT=YesPlease gettext support will be #defined away for C
programs. For Shell and Perl programs we rely on the git message
catalog not being avalalable. That's a reasonable assumption since the
*.po files won't be installed if NO_GETTEXT=YesPlease is specified.

The gettext wrappers that are provided in the patch are only the bare
minimum required to begin translation work. In particular I haven't
added wrappers for the gettext functions that enable plural support,
or those that provide message context (msgctxt).

All of that can be added later. The intent is to start with a small
subset and see what we need later, not to start with something that's
unnecessarily large right away.

Gettext .mo files will be installed and looked for in
$(prefix)/share/locale. GIT_LOCALEDIR can also be set to override
that, this variable is only intended for testing Git itself.

This patch is based on work by Jeff Epler <jepler@unpythonic.net> who
did the initial Makefile / C work.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>