Documentation: Add variable-substitution script
The new subst-config.perl is run by the Makefile as

  ./subst-config.perl <vars> <in> <out>

It first reads <vars> for variable-documenting blocks.  A line
consisting of '<name>::' starts the block for <name>, except inside a
'--' delimited block (which is used to generate sub-lists).  A block
extends until the first blank line, which is what asciidoc also looks
for.

Then it copies from <in> to <out>, substituting lines of the form

  @@CONFIG(<name>)@@

with the documentation block for <name>.

Signed-off-by: Thomas Rast <trast@student.ethz.ch>