Documentation/Makefile: Specify -f user-manual.conf
The user-manual.xml target was broken in the "Add
variable-substitution script" commit.

It broke because asciidoc(1) will implicitly search a configuration
file alongsite the .txt file it's generating, but that didn't work
with the preprocessed .txt.cv file (which would need a
user-manual.txt.conf file)

Fix that by specifying the path to the config file explicitly like the
other asciidoc invocations do.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>